rename: 폴더 구조 변경 및 이동
