+++ src/components/TransactionScreen/FinancialCalendar.tsx
@@ -1,7 +1,7 @@
 import {View, StyleSheet} from 'react-native';
 import {Calendar} from 'react-native-big-calendar';
 import {theme} from 'styles';
-import {calendarTheme} from 'constants/TransactionScreen';
+import {calendarTheme} from 'constants/screens/TransactionScreen';
 import DayHeader from 'components/TransactionScreen/DayHeader';
 
 const dummyEvents = [
