+++ src/constants/components/Transaction.ts
@@ -1,0 +1,5 @@
+export const categoryList = {
+  profit: '수익',
+  cost: '비용',
+  account: '계정',
+} as const;
