+++ src/navigators/constants/icon.ts
@@ -1,5 +1,5 @@
 import {TabScreenName} from 'navigators/types';
-import {KeyOfIcons} from 'types/icon';
+import {KeyOfIcons} from 'types/components/icon';
 
 export const TabNavigatorIcon: Record<TabScreenName, KeyOfIcons> = {
   Ledger: 'MoneyOut',
