+++ src/types/components/Transaction.ts
@@ -1,0 +1,4 @@
+export type DebitCreditType = {
+  name: string;
+  amount: string;
+};
