+++ src/types/components/funnel.ts
